Separate cursor loading from value history
