+++ differential-dataflow/src/operators/cursor/mod.rs
@@ -1,4 +1,6 @@
 //! Operator implementations based on cursor traversal.
 
+mod history;
+
 pub mod join;
 pub mod reduce;
