+++ differential-dataflow/src/operators/int_proxy/history.rs
@@ -1,5 +1,4 @@
 //! Time-ordered replay of proxy update histories, with meet-advancement.
 
 /// A value history suitable for integer proxy values.
-pub(in crate::operators) type IdHistory<T, R> = crate::operators::ValueHistory<u64, T, R>;
-
+pub(in crate::operators) type IdHistory<T, R> = crate::operators::history::ValueHistory<u64, T, R>;
